closes #733
