+++ srcv2/pages/CreateProjectPage.js
@@ -1,27 +1,47 @@
 import React, { useState } from 'react'
 import { useMutation } from '@apollo/client'
+import { useHistory } from 'react-router-dom'
 
 import CreateProject from '../components/CreateProject'
+import CreateProjectFunding from '../components/CreateProjectFunding'
 import Section from '../components/Section'
 
-import { ADD_PROJECT } from '../operations/mutations/ProjectMutations'
+import { ADD_PROJECT, UPDATE_PROJECT } from '../operations/mutations/ProjectMutations'
 import { CREATE_PERMISSION } from '../operations/mutations/PermissionMutations'
 
 import { sessionUser } from '../reactivities/variables'
+
+import { CURRENCIES, FUNDING_PLAN_TIMEFRAME_AMOUNTS } from '../constants'
 
 const CreateProjectPage = () => {
 
     const [selectedUser, setSelectedUser] = useState(null)
     const [selectedRepo, setSelectedRepo] = useState(null)
+    const [budgetRange, setBudgetRange] = useState(null)
+    const [currency, setCurrency] = useState(CURRENCIES[0])
+    const [timeframeAmount, setTimeFrameAmount] = useState(FUNDING_PLAN_TIMEFRAME_AMOUNTS[0])
+
+    const history = useHistory()
 
     const [
         addProject,
         {
-            data,
+            data: addProjectData,
             loading: loadingAddProject,
-            error
+            error: errorProjectData
         }
     ] = useMutation(ADD_PROJECT, {
+        errorPolicy: 'all'
+    })
+
+    const [
+        addProjectFunding,
+        {
+            data,
+            loading,
+            error
+        }
+    ] = useMutation(UPDATE_PROJECT, {
         errorPolicy: 'all'
     })
 
@@ -71,8 +91,16 @@ const CreateProjectPage = () => {
     const saveAndContinue = async () => {
         try {
             const newProject = await createProject()
-            setProjectCreated(newProject)
-            goToNextSection()
+            const fundProjectVariables = {
+                project_id: newProject.data.createProject.id,
+                expected_budget: budgetRange,
+                expected_budget_timeframe: timeframeAmount,
+                expected_budget_currency: currency.name
+            }
+            await addProjectFunding({ 
+                variables: fundProjectVariables
+            })
+            history.push('/dashboard')
         } catch (err) {
             console.log(err)
         }
@@ -80,12 +108,19 @@ const CreateProjectPage = () => {
 
     return (
         <div className='CreateProject'>
-            Create Project
             <CreateProject
                 selectedUser={selectedUser}
                 selectedRepo={selectedRepo}
                 setSelectedUser={setSelectedUser}
                 setSelectedRepo={setSelectedRepo}
+            />
+            <CreateProjectFunding
+                budgetRange={budgetRange}
+                currency={currency}
+                timeframeAmount={timeframeAmount}
+                setBudgetRange={setBudgetRange}
+                setCurrency={setCurrency}
+                setTimeFrameAmount={setTimeFrameAmount}
             />
             <Section>
                 <button
